ComplexDataTypeDescriptor --> COmplexDataFormat
Override object equals method
